requirements.txt
# Python runtime
python_version==3.12

# Web + DB
Flask>=2.2
PyMySQL>=1.0.2
PyYAML>=6.0
python-dotenv>=1.0.0

# Data science
pandas>=2.2
numpy>=1.26
statsmodels>=0.14
scikit-learn>=1.3
lightgbm>=4.0
xgboost>=2.1
joblib>=1.3

# Monitoring & utils
prometheus-client>=0.17
loguru>=0.7.0
requests>=2.31

# Optional: for typed config
pydantic>=2.0
